feat: split server/core into subpackages and fix tests
