+++ packages/solid/package.json
@@ -21,17 +21,23 @@
   "files": [
     "dist"
   ],
-  "source": "index.tsx",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
+  "main": "dist/cjs/index.module.js",
+  "module": "dist/core/index.module.js",
+  "types": "dist/core/core.d.ts",
   "exports": {
     ".": {
-      "solid": "./dist/source/index.jsx",
-      "import": "./dist/esm/index.js",
-      "browser": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js",
-      "node": "./dist/cjs/index.js"
+      "solid": "./dist/core/core.jsx",
+      "import": "./dist/core/index.module.js",
+      "browser": "./dist/core/index.module.js",
+      "require": "./dist/core/index.module.js",
+      "node": "./dist/core/index.module.js"
+    },
+    "./server": {
+      "solid": "./dist/server/server.jsx",
+      "import": "./dist/server/index.module.js",
+      "browser": "./dist/server/index.module.js",
+      "require": "./dist/server/index.common.js",
+      "node": "./dist/server/index.common.js"
     }
   },
   "sideEffects": false,
@@ -53,9 +59,9 @@
     "babel-preset-solid": "^0.24.13",
     "jiti": "^1.6.4",
     "rollup": "^2.42.0",
-    "rollup-preset-solid": "^0.1.2",
+    "rollup-preset-solid": "^0.3.0",
     "solid-js": "^0.24.15",
-    "twind": "^0.16.4",
+    "twind": "0.16.3",
     "typescript": "^4.2.3"
   },
   "scripts": {
