+++ packages/solid/rollup.config.js
@@ -1,3 +1,11 @@
 import withSolid from 'rollup-preset-solid'
 
-export default withSolid({ external: [/twind\/*/] })
+export default withSolid([
+  { input: 'src/core.tsx', external: [/twind\/*/] },
+  {
+    input: 'src/server.tsx',
+    external: [/twind\/*/],
+    targets: ['cjs', 'esm'],
+    writePackageJson: true,
+  },
+])
